Bug Fixes
- Delete buggy error window
- Fix redundant conditional in settings menu
- Fix Item Settings show up when there are no valid commands

+++ Source/RimConnection/API/RimConnectAPI.cs
@@ -2,7 +2,6 @@
 using Verse;
 using RestSharp;
 using System;
-using RimConnection.Windows;
 
 namespace RimConnection
 {
@@ -63,7 +62,7 @@ namespace RimConnection
                 if (validCommandResponse.StatusCode != System.Net.HttpStatusCode.OK)
                 {
                     RimConnectSettings.initialiseSuccessful = false;
-                    BugReport.CreateBugReport("Failed to provide valid commands to the server");
+                    Log.Error("Failed to provide valid commands to the server");
                 }
                 else
                 {
@@ -78,8 +77,7 @@ namespace RimConnection
             }
             catch(Exception e)
             {
-                Log.Warning(e.Message);
-                BugReport.CreateBugReport($"Failed to provide valid commands to server. {e.Message}");
+                Log.Error($"Failed to provide valid commands to server. {e.Message}");
             }
         }
 
@@ -98,8 +96,7 @@ namespace RimConnection
 
                 if (validCommandResponse.StatusCode != System.Net.HttpStatusCode.OK)
                 {
-                    Log.Warning("Failed to update command options to server");
-                    BugReport.CreateBugReport($"Failed to update command options to server. {validCommandResponse.StatusCode}");
+                    Log.Error($"Failed to update command options to server. {validCommandResponse.StatusCode}");
                 }
                 else
                 {
@@ -108,8 +105,7 @@ namespace RimConnection
             }
             catch (Exception e)
             {
-                Log.Warning(e.Message);
-                BugReport.CreateBugReport($"Failed to update command options to server. {e.Message}");
+                Log.Error($"Failed to update command options to server. {e.Message}");
             }
         }
 
@@ -122,13 +118,15 @@ namespace RimConnection
             try
             {
                 var response = client.Execute<CommandList>(baseRequest);
+                if (response == null) throw new NullReferenceException("Response is null");
                 var commands = response.Data.commands;
+               
 
                 DeleteCommands(commands.Count);
                 return commands;
             } catch(Exception e)
             {
-                BugReport.CreateBugReport($"Failed to get commands from server. {e.Message}");
+                Log.Error($"Failed to get commands from server. {e.Message}");
                 throw;
             }
 
@@ -153,7 +151,7 @@ namespace RimConnection
                 var response = client.Execute(baseRequest);
             } catch (Exception e)
             {
-                BugReport.CreateBugReport($"Failed to delete commands from server. {e.Message}");
+                Log.Error($"Failed to delete commands from server. {e.Message}");
                 throw;
             }
         }
@@ -169,7 +167,7 @@ namespace RimConnection
                 var response = client.Execute(baseRequest);
             } catch(Exception e)
             {
-                BugReport.CreateBugReport($"Failed to provide world info to server. {e.Message}");
+                Log.Error($"Failed to provide world info to server. {e.Message}");
                 throw;
             }
         }
